Refactor page handling in DgusTFT class
The way that the DgusTFT::IdleLoop() function handled page indices was refactored from a series of if-else statements to a switch-case statement, which improves code readability and maintainability. Additionally, debugging code was added to help identify when an invalid page index is encountered. Some handler function names were also modified for consistency and some unused functions were removed. This refactoring does not alter the way the code works but improves its organization and ease of comprehension.